Move models to a directory, update imports
